requirements.txt
accelerate
bitsandbytes
datasets
deepspeed==0.14.4
einops
flash-attn==2.6.1
isort
jsonlines
loralib
optimum
packaging
peft
ray[default]==2.12.0
torchmetrics
tqdm
transformers==4.42.4
transformers_stream_generator
wandb
wheel
